fastAPI/fastapi_code2/ddd/pyproject.toml
[project]
name = "ddd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.2",
    "fastapi>=0.119.1",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "tomlkit>=0.13.3",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.38.0",
]
